requirements.txt
pyautogen
chainlit
matplotlib
numpy
pyvista
gmsh
meshio
h5py